DesktopOrganizer/requirements.txt
watchdog>=4.0.0
pdfminer.six==20221105
python-docx==0.8.11
colorama==0.4.6
python-dotenv==1.0.0
google-genai